Add mathD

+++ Pro.100.Game.Ua/src/Discrete/Math/Function/Class/DiscreteMathDecussation.java
@@ -14,7 +14,7 @@ public class DiscreteMathDecussation implements DiscreteMathFunction {
 		Iterator iteratorA = A.iterator();
 
 		while (iteratorA.hasNext()) {
-			Integer integerA = iteratorA.next();
+			Integer integerA = (Integer) iteratorA.next();
 			Iterator iteratorB = B.iterator();
 			
 			while (iteratorB.hasNext()) {
